requirements.txt
numpy>=1.26
scipy>=1.11
scikit-image>=0.22
PyWavelets>=1.5
matplotlib>=3.8
Pillow>=10.0
pytest>=8.0
